test: setup c unit tests
