+++ .gitignore
@@ -2,3 +2,4 @@
 iso-chooser-menu
 out.vars
 *.json
+runtests
